feature(project): Open window and display shader output
The current state of the project is very messy but it is now possible to open a window and display the output of a shader to it. This was done by creating a storage image and rendering to it and then copying the result to the swapchain before presenting it.