refactor(project): Optimize the rendering process
Remove unecessary data sent to the GPU (aka. set the range value to the y value of the pos_zw vector); and also change the range to just be one f32. Rename some and remove unused/unecessary structs and properties in the shader. Change the renderer to use the slabs collision detection for a ray (this is much faster than the previous version, but requires a 16/17-level nested for loops).